Homepage blocks ready
Changing color SVG for block "Our values"
-----------------------------------------

SVG is supposed to change colors, on title, icon and hover
Add composer package to include uploaded svg
Modify SVG to have customizable color (path fill="current")
Adding scratches to bottom

Block what we do
----------------

This is a tricky one because flow has to go vertical
and there's no way of doing that with native css

This one also has background that follows line breaks
this isn't possible with css, to make some tricks to achieve same effect

Block Call to action
--------------------

This block is global, customizable in options

Block Hero
----------

At first, this image was part of page template, with text as part of image.

Change it to a gutenberg block to make it easier to use, with customizable text.

Ready for review
----------------

With this, homepage is ready for review